Move frontend error handling to __init__, handle AuthorizationError
